refactor(tools): move availability_view out of orchestration layer
Extract integration source normalization to app/tools/utils so runtime
execution no longer imports orchestration nodes. Also set connection_verified
when _grafana_source unwraps classified Pydantic configs.
