+++ app/core/orchestration/node/investigate/tools.py
@@ -12,6 +12,7 @@ from app.core.runtime import public_tool_input
 from app.services.agent_llm_client import ToolCall
 from app.tools.registered_tool import RegisteredTool
 from app.tools.registry import get_registered_tools
+from app.tools.utils.integration_sources import availability_view
 from app.utils.tool_trace import redact_sensitive
 
 # Consecutive iterations made up ENTIRELY of duplicate (already-seen) tool calls
@@ -32,28 +33,6 @@ STAGNATION_NUDGE = (
 def get_available_tools(resolved_integrations: dict[str, Any]) -> list[RegisteredTool]:
     available_sources = availability_view(resolved_integrations)
     return [t for t in get_registered_tools("investigation") if t.is_available(available_sources)]
-
-
-def availability_view(resolved_integrations: dict[str, Any]) -> dict[str, Any]:
-    """Adapt resolved integration configs to the legacy tool availability contract."""
-    from pydantic import BaseModel
-
-    view: dict[str, Any] = {}
-    for key, value in resolved_integrations.items():
-        if key.startswith("_"):
-            view[key] = value
-            continue
-        if isinstance(value, BaseModel):
-            item = value.model_dump(exclude_none=True)
-            item.setdefault("connection_verified", True)
-            view[key] = item
-        elif isinstance(value, dict) and value:
-            item = dict(value)
-            item.setdefault("connection_verified", True)
-            view[key] = item
-        else:
-            view[key] = value
-    return view
 
 
 def build_connected_tool_context(
